// agents/skills/distill-page/scripts/proto/common_quality_data_pb.d.ts
export declare type AnnotatedPageContent = Message<"optimization_guide.proto.AnnotatedPageContent"> & {
  /**
   * The version of the proto. Each time a backwards incompatible change is made
   * to the proto, the version should be incremented.
   *
   * @generated from field: optimization_guide.proto.AnnotatedPageContentVersion version = 1 [features.field_presence = EXPLICIT];
   */
  version: AnnotatedPageContentVersion;

  /**
   * The mode used to extract this proto. The data in the proto depends on the
   * mode. See comments to understand which data is available for each mode.
   *
   * @generated from field: optimization_guide.proto.AnnotatedPageContentMode mode = 6 [features.field_presence = EXPLICIT];
   */
  mode: AnnotatedPageContentMode;

  /**
   * The root node of the ContentNode tree.
   *
   * @generated from field: optimization_guide.proto.ContentNode root_node = 2 [features.field_presence = EXPLICIT];
   */
  rootNode?: ContentNode | undefined;

  /**
   * The data for the main frame of the page.
   *
   * @generated from field: optimization_guide.proto.FrameData main_frame_data = 3 [features.field_presence = EXPLICIT];
   */
  mainFrameData?: FrameData | undefined;

  /**
   * The interaction information for the page--the subtree represented by the
   * AnnotatedPageContent. This is for top-level interactions, such as the mouse
   * position and focused element, for which there can only be one unique value
   * for the entire page.
   * Currently this is only tracked for the main frame.
   *
   * @generated from field: optimization_guide.proto.PageInteractionInfo page_interaction_info = 4 [features.field_presence = EXPLICIT];
   */
  pageInteractionInfo?: PageInteractionInfo | undefined;

  /**
   * The rect of the viewport for the tab. i.e., screenshot dimensions.
   * Dimensions are in captured pixels.
   * This coordinate system should be used for coordinate based browser actions.
   * This uses device independent pixels, i.e., doesn't include the device
   * scale factor while other dimensions in APC are based on that.
   *
   * @generated from field: optimization_guide.proto.BoundingRect viewport_geometry = 5 [features.field_presence = EXPLICIT];
   */
  viewportGeometry?: BoundingRect | undefined;

  /**
   * The id of the tab that the page content was extracted from. Guaranteed to
   * be unique within the browsing session.
   *
   * @generated from field: int32 tab_id = 7 [features.field_presence = EXPLICIT];
   */
  tabId: number;

  /**
   * Provides APC for internal popup widgets created by elements like <select>
   * and <input type="color"> . Does not provide APC for separate windows opened
   * via window.open() or similar. A page can only have one popup window open at
   * a time. If one is open, its details are populated here. The popup window
   * draws on top of the tab and as such has a higher z order than the main
   * frame's Document.
   *
   * @generated from field: optimization_guide.proto.PopupWindow popup_window = 8 [features.field_presence = EXPLICIT];
   */
  popupWindow?: PopupWindow | undefined;

  /**
   * Information that pertains to the user's profile, not the page.
   * TODO(crbug.com/454610456): This should be moved out of AnnotatedPageContent because
   * it does not semantically belong here.
   *
   * @generated from field: optimization_guide.proto.ProfileInformation profile_information = 9 [features.field_presence = EXPLICIT];
   */
  profileInformation?: ProfileInformation | undefined;

  /**
   * Metadata about the page when Gemini in Chrome captures Annotated Page
   * Content.
   * TODO(crbug.com/484011242): This should be moved out of AnnotatedPageContent because
   * it does not semantically belong here.
   *
   * @generated from field: optimization_guide.proto.GeminiInChromePageMetadata gemini_in_chrome_page_metadata = 10 [features.field_presence = EXPLICIT];
   */
  geminiInChromePageMetadata?: GeminiInChromePageMetadata | undefined;
};

/**
 * Describes the message optimization_guide.proto.AnnotatedPageContent.
 * Use `create(AnnotatedPageContentSchema)` to create a new message.
 */
export declare const AnnotatedPageContentSchema: GenMessage<AnnotatedPageContent>;
// ...
export declare type ContentNode = Message<"optimization_guide.proto.ContentNode"> & {
  /**
   * Note: CONTENT_ATTRIBUTE_TEXT and CONTENT_ATTRIBUTE_IMAGE are leaf nodes
   * which don't have any children nodes.
   *
   * @generated from field: repeated optimization_guide.proto.ContentNode children_nodes = 1;
   */
  childrenNodes: ContentNode[];

  /**
   * @generated from field: optimization_guide.proto.ContentAttributes content_attributes = 2 [features.field_presence = EXPLICIT];
   */
  contentAttributes?: ContentAttributes | undefined;
};
// ...
export declare type ProfileInformation = Message<"optimization_guide.proto.ProfileInformation"> & {
  /**
   * @generated from field: optimization_guide.proto.AutofillInformation autofill_information = 1 [features.field_presence = EXPLICIT];
   */
  autofillInformation?: AutofillInformation | undefined;
};
// ...
export enum AnnotatedPageContentVersion {
  /**
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_VERSION_UNKNOWN = 0;
   */
  ANNOTATED_PAGE_CONTENT_VERSION_UNKNOWN = 0,

  /**
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_VERSION_1_0 = 1;
   */
  ANNOTATED_PAGE_CONTENT_VERSION_1_0 = 1,

  /**
   * DEPRECATED: Use AnnotatedPageContentMode mode instead.
   * TODO(crbug.com/490195093): Maintain both on the client until all users have been
   * switched over.
   *
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_VERSION_ONLY_ACTIONABLE_ELEMENTS_1_0 = 2 [deprecated = true];
   * @deprecated
   */
  ANNOTATED_PAGE_CONTENT_VERSION_ONLY_ACTIONABLE_ELEMENTS_1_0 = 2,

  /**
   * The maximum valid version of the proto. This value includes deprecated
   * versions to ensure backward compatibility with older clients.
   *
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_VERSION_MAX_VALID = 2;
   */
  ANNOTATED_PAGE_CONTENT_VERSION_MAX_VALID = 2,
}

/**
 * Describes the enum optimization_guide.proto.AnnotatedPageContentVersion.
 */
export declare const AnnotatedPageContentVersionSchema: GenEnum<AnnotatedPageContentVersion>;

/**
 * The extraction on the client can be configured to a specific mode. This
 * tracks the mode used to extract this proto.
 *
 * @generated from enum optimization_guide.proto.AnnotatedPageContentMode
 */
export enum AnnotatedPageContentMode {
  /**
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_MODE_DEFAULT = 0;
   */
  DEFAULT = 0,

  /**
   * This mode is strictly a superset of the default mode. It includes
   * additional information to understand which content a user can interact
   * with.
   *
   * Note: This mode is under active development and backwards incompatible
   * changes are expected.
   *
   * @generated from enum value: ANNOTATED_PAGE_CONTENT_MODE_ACTIONABLE_ELEMENTS = 1;
   */
  ACTIONABLE_ELEMENTS = 1,
}
// ...
export enum ContentAttributeType {
  /**
   * Unknown is not a valid attribute type and should never be used by the
   * client. You can consider discarding ContentNodes with this attribute type.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_UNKNOWN = 0;
   */
  CONTENT_ATTRIBUTE_UNKNOWN = 0,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_ROOT = 1;
   */
  CONTENT_ATTRIBUTE_ROOT = 1,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_CONTAINER = 2;
   */
  CONTENT_ATTRIBUTE_CONTAINER = 2,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_IFRAME = 3;
   */
  CONTENT_ATTRIBUTE_IFRAME = 3,

  /**
   * Text chunks.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_PARAGRAPH = 4;
   */
  CONTENT_ATTRIBUTE_PARAGRAPH = 4,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_HEADING = 5;
   */
  CONTENT_ATTRIBUTE_HEADING = 5,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_TEXT = 20;
   */
  CONTENT_ATTRIBUTE_TEXT = 20,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_ANCHOR = 22;
   */
  CONTENT_ATTRIBUTE_ANCHOR = 22,

  /**
   * Image chunks.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_IMAGE = 9;
   */
  CONTENT_ATTRIBUTE_IMAGE = 9,

  /**
   * Root SVG nodes.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_SVG_ROOT = 25;
   */
  CONTENT_ATTRIBUTE_SVG_ROOT = 25,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_CANVAS = 26;
   */
  CONTENT_ATTRIBUTE_CANVAS = 26,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_VIDEO = 27;
   */
  CONTENT_ATTRIBUTE_VIDEO = 27,

  /**
   * Structured chunks.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_ORDERED_LIST = 6;
   */
  CONTENT_ATTRIBUTE_ORDERED_LIST = 6,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_UNORDERED_LIST = 7;
   */
  CONTENT_ATTRIBUTE_UNORDERED_LIST = 7,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_LIST_ITEM = 23;
   */
  CONTENT_ATTRIBUTE_LIST_ITEM = 23,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_FORM = 8;
   */
  CONTENT_ATTRIBUTE_FORM = 8,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_FORM_CONTROL = 24;
   */
  CONTENT_ATTRIBUTE_FORM_CONTROL = 24,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_TABLE = 10;
   */
  CONTENT_ATTRIBUTE_TABLE = 10,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_TABLE_ROW = 21;
   */
  CONTENT_ATTRIBUTE_TABLE_ROW = 21,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_TABLE_CELL = 11;
   */
  CONTENT_ATTRIBUTE_TABLE_CELL = 11,

  /**
   * Dialog/modal containers.
   *
   * @generated from enum value: CONTENT_ATTRIBUTE_DIALOG_MODAL = 28;
   */
  CONTENT_ATTRIBUTE_DIALOG_MODAL = 28,

  /**
   * @generated from enum value: CONTENT_ATTRIBUTE_DIALOG_MODELESS = 29;
   */
  CONTENT_ATTRIBUTE_DIALOG_MODELESS = 29,
}
// ...
export enum TextSize {
  /**
   * @generated from enum value: TEXT_SIZE_M_DEFAULT = 0;
   */
  M_DEFAULT = 0,

  /**
   * @generated from enum value: TEXT_SIZE_XS = 1;
   */
  XS = 1,

  /**
   * @generated from enum value: TEXT_SIZE_S = 2;
   */
  S = 2,

  /**
   * @generated from enum value: TEXT_SIZE_L = 3;
   */
  L = 3,

  /**
   * @generated from enum value: TEXT_SIZE_XL = 4;
   */
  XL = 4,
}
// ...
export enum TableRowType {
  /**
   * @generated from enum value: TABLE_ROW_TYPE_UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * @generated from enum value: TABLE_ROW_TYPE_HEADER = 1;
   */
  HEADER = 1,

  /**
   * @generated from enum value: TABLE_ROW_TYPE_BODY = 2;
   */
  BODY = 2,

  /**
   * @generated from enum value: TABLE_ROW_TYPE_FOOTER = 3;
   */
  FOOTER = 3,
}
// ...
export enum AnnotatedRole {
  /**
   * @generated from enum value: ANNOTATED_ROLE_UNKNOWN = 0;
   */
  UNKNOWN = 0,

  /**
   * Landmark/section chunks.
   *
   * @generated from enum value: ANNOTATED_ROLE_HEADER = 5;
   */
  HEADER = 5,

  /**
   * @generated from enum value: ANNOTATED_ROLE_NAV = 6;
   */
  NAV = 6,

  /**
   * @generated from enum value: ANNOTATED_ROLE_SEARCH = 7;
   */
  SEARCH = 7,

  /**
   * @generated from enum value: ANNOTATED_ROLE_MAIN = 8;
   */
  MAIN = 8,

  /**
   * @generated from enum value: ANNOTATED_ROLE_ARTICLE = 9;
   */
  ARTICLE = 9,

  /**
   * @generated from enum value: ANNOTATED_ROLE_SECTION = 10;
   */
  SECTION = 10,

  /**
   * @generated from enum value: ANNOTATED_ROLE_ASIDE = 11;
   */
  ASIDE = 11,

  /**
   * @generated from enum value: ANNOTATED_ROLE_FOOTER = 12;
   */
  FOOTER = 12,

  /**
   * Indicates that the subtree for this ContentNode is hidden from the user but
   * accessible via features like find-in-page. See `hidden=until-found` in
   * https://developer.mozilla.org/en-US/docs/Wecrbug.com/HTML/Global_attributes/hidden
   * for details.
   *
   * Note: This is different from Geometry.visible_bounding_box being empty
   * which indicates whether the content is offscreen. The visible_bounding_box
   * for nodes in a hidden subtree may be non-empty, depending on whether the
   * content would be in the viewport if the user searched for it.
   *
   * @generated from enum value: ANNOTATED_ROLE_CONTENT_HIDDEN = 13;
   */
  CONTENT_HIDDEN = 13,

  /**
   * Indicates that the subtree for this ContentNode contains content marked as
   * isAccessibleForFree=false by a schema.org CreativeWork declaration, as
   * described at
   * https://developers.google.com/search/docs/appearance/structured-data/paywalled-content.
   *
   * @generated from enum value: ANNOTATED_ROLE_PAID_CONTENT = 14;
   */
  PAID_CONTENT = 14,
}